+++ pygbe/main.py
@@ -23,7 +23,7 @@ from pygbe.class_initialization import initialize_surface, initialize_field
 from pygbe.output import print_summary
 from pygbe.matrixfree import (generateRHS, generateRHS_gpu, calculate_solvation_energy,
                               coulomb_energy, calculate_surface_energy)
-from pygbe.util.readData import readParameters
+from pygbe.util.read_data import readParameters
 from pygbe.tree.FMMutils import computeIndices, precomputeTerms, generateList
 
 try:
